refactor(website): file name convention
